// src/app/models/re.ts
export class Re {
  title: string;
  imageUrl: string;
  address: string;
  city: string;
  contact: string;
  price: number;
  category: string;
  floor: number;
  area: number;
  rooms: number;
  bathrooms: number;
  totalFloors: number;
  elevator: boolean;
  balcony: boolean;
  description: string;

  
}